+++ CHANGES.rst
@@ -1,6 +1,8 @@
 4.4.2 (unreleased)
 ==================
 
+- Fixes traceback raised by the stretch histogram in plot options when entering multiselect mode. [#3825]
+
 Bug Fixes
 ---------